rename auth_service to user_service
