+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:sync_play/app.dart';
 import 'package:sync_play/services/app_service.dart';
-import 'package:sync_play/services/auth_service.dart';
+import 'package:sync_play/services/user_service.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
