+++ lib/ui/pages/register/register_bloc.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:sync_play/models/app_binding.dart';
 import 'package:sync_play/models/app_error.dart';
-import 'package:sync_play/services/auth_service.dart';
+import 'package:sync_play/services/user_service.dart';
 import 'package:sync_play/util/util.dart';
 
 class RegisterPageBindings implements AppBinding {
